Move 'module' sub-switches to pages
