+++ packages/ui/src/app/pages/Forum/index.ts
@@ -1,1 +1,1 @@
-export * from './Forum'
+export * from './ForumModule'
